Add a control to disable processing sines in the driver thread
A single busy thread causes non-work-interval threads to process on a big core
for all buffer sizes and without any artificial load in real-time threads. To
avoid using the private work interval API, we can use the driver thread only
to wake up and wait for workers to complete, but not to process sines. Add a
control to switch between these modes.

The sine bank contains an unused StereoBuffer when ProcessInDriverThread is
false. I don't think it's worth additional logic to avoid this.