Set the thread policy of worker threads after changing the buffer size
The thread policy of worker threads wasn't updated after changing the buffer
size, so the period and constraint parameters would be wrong.

It would be possible to call setThreadTimeConstraintPolicy() for all worker
threads to avoid stopping the driver, but it seems more realistic to keep the
policy constant. The approach of stopping and starting the driver is needed for
turning on the work interval either way, and doing so is fast.