Get rid of the missing time color functionality in ActivityView
missingTimeColor was set for the load graph so that a stream of dropouts is
shown as one large red area instead of alternating between a red line for the
overshot deadline and a blank space for the subsequent skipped buffer.

Adding missing time points with value 1.0 is problematic because if subsequent
values are added within the same integer point, the value stays at 1.0 but
the color is that of the added value. This results in spikes that don't exist
being drawn after missing time regions.

Having filled drop-out regions doesn't seem worth the complexity of fixing this,
so I'm removing it.