Switch to using SCHED_OTHER for the busy thread's scheduling policy
According to the POSIX standard, "this policy [SCHED_OTHER] is defined to
allow strictly conforming applications to be able to indicate in a portable
manner that they no longer need a realtime scheduling policy". In XNU,
SCHED_OTHER uses the timesharing policy, which I believe has lower priority than
FIFO and RR. In any case, I haven't seen a difference in audio performance
after switching.

This is also consistent with the policy used in RealtimeLib's
setIdleThreadPriority() and consistency is important when testing.