Adjust the default busy thread settings to avoid background termination
With the old settings, the app would be terminated within 15 seconds when
running in the background with audio input enabled due to the busy thread
exceeding the CPU usage limit. With the new settings, the app can run
indefinitely as long as audio threads stay under the limit as well.

Increasing the busy thread period allows the CPU usage to be lowered without
changing the transient response or energy usage.