Remove Driver::setIsInputEnabled()
In preparation for adding Driver::workgroup(), which can change after creating a
new AURemoteIO instance when enabling/disabling the audio input. It would be
fragile for the previously returned workgroup to become invalid after changing
some Driver configuration.

We now recreate the Driver in full when enabling/disabling the input. Worker
threads are also destroyed and recreated in preparation for joining the
workgroup of the new Driver instance.