Rename mNumRequestedWorkerThreads -> mNumWorkerThreads
It was named "requested" in d30b167122 to distinguish it from the return value
of numWorkerThreads(), which uses the size of the mWorkerThreads vector. This
will change later on the branch, at which point it will be confusing for the
member name to not simply match the setter/getter name.