Replace worker thread counts with processing thread counts
In preparation for adding a special value to use the recommended number of
processing threads in AudioHostConfig.

When using the number of worker threads, kStandardPerformanceConfig somehow
needs to request "RecommendedNumProcessingThreads - 1" worker threads and
kOptimalPerformanceConfig needs to request "RecommendedNumProcessingThreads"
worker threads. Expressing configs using the number of processing threads avoids
the need for two special values.

Also make the use of the sine bank in Engine more explicit by not configuring a
thread that is sometimes unused.